requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
apturl==0.5.2
asgiref==3.6.0
async-timeout==4.0.3
attrs==22.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.15.1
channels==4.0.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==40.0.1
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.1
Django==4.1.5
django-bootstrap5==23.3
django-filter==23.3
django-nextjs==2.4.0
djangorestframework==3.14.0
dnspython==2.3.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
Flask==2.2.3
frozenlist==1.4.0
future==0.18.2
httplib2==0.20.2
idna==3.3
impacket==0.10.0
importlib-metadata==4.6.4
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jsonschema==4.17.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.9.1
ldapdomaindump==0.9.4
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
MarkupSafe==2.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
netifaces==0.11.0
numpy==1.26.1
oauthlib==3.2.0
olefile==0.46
opencv-python==4.8.1.78
paramiko==2.9.3
pefile==2023.2.7
pexpect==4.8.0
Pillow==9.0.1
pipenv==2022.12.19
platformdirs==2.5.1
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.17
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymssql==2.2.7
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==2.4.7
pypng==0.20220715.0
pyproject-toml==0.0.10
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.4.2
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
speedtest-cli==2.1.3
sqlparse==0.4.3
systemd-python==234
texttable==1.6.4
toml==0.10.2
typing_extensions==4.5.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
wadllib==1.3.6
websocket-client==1.2.3
websockets==11.0.3
Werkzeug==2.2.3
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
